% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricT.R
\name{mt_null_plot}
\alias{mt_null_plot}
\title{Permutation null distribution of T for one condition}
\usage{
mt_null_plot(x, condition = 1L)
}
\arguments{
\item{x}{A \code{metricT} object computed with \code{keep_null = TRUE}.}

\item{condition}{Row number of \code{x$table}.}
}
\description{
Permutation null distribution of T for one condition
}
