% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{run_metricT_app}
\alias{run_metricT_app}
\title{Launch the Metric-T graphical interface}
\usage{
run_metricT_app(...)
}
\arguments{
\item{...}{Passed to \code{shiny::runApp}.}
}
\value{
No return value, called for side effects (starts the 'shiny' application).
}
\description{
Launch the Metric-T graphical interface
}
\examples{
if (interactive()) {
  run_metricT_app()
}
}
